requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
gunicorn==21.2.0
cryptography==41.0.7
